+++ openhands/events/serialization/action.py
@@ -69,7 +69,7 @@ def action_from_dict(action: dict) -> Action:
     # images_urls has been renamed to image_urls
     if 'images_urls' in args:
         args['image_urls'] = args.pop('images_urls')
-        
+
     try:
         decoded_action = action_class(**args)
         if 'timeout' in action:
